agent/requirements.txt
psutil
requests
python-dotenv
websocket-client